+++ test/qunit/modules/preload/good.js
@@ -1,0 +1,7 @@
+(function (require) {
+    ok(true, "Should be initialized once");
+
+    return {
+        "ok": true
+    };
+})